Pull some of the properties variables into the files that use them
